add callback object for filtering callback handlers
