+++ telegram_bot/src/callback_data_objects.py
@@ -1,0 +1,5 @@
+from aiogram.utils.callback_data import CallbackData
+
+from constants import CALLBACK_QUERY_TYPE
+
+BRAND_CALLBACK_OBJECT = CallbackData(CALLBACK_QUERY_TYPE, 'data')  # brand:<data>
