+++ telegram_bot/src/constants.py
@@ -1,3 +1,6 @@
 # components
 CONFIRM_BUTTON_TEXT: str = 'OK'
 CATALOGUE_CHECKBOX_EMOJI: str = ':ballot_box_with_check:'
+
+# callback query types
+CALLBACK_QUERY_TYPE: str = 'brand'
