+++ telegram_bot/src/handlers/__init__.py
@@ -1,0 +1,1 @@
+from . import start_form
